requirements.txt
allure-pytest
pytest
pytest-xdist
requests
selenium
